fix(repo): ensure options only accepts one of html or url but not both
refactor(repo): move urlboxoptions tests to own file
