+++ Urlbox.MsTest/Urlbox.MsTest.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <!-- <TargetFramework>netcoreapp2.0</TargetFramework> -->
     <TargetFramework>net6.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
